Add files via upload
This is a Computation Math Project with using MATLAB as the software. This cryptoghraphy is proposed to convert a data or image to another data or image. In the process, the data or image is converted to a numerical data, in this case, matrix. Then, the matrix is modified with some numerical numbers or can be inversed, depend to the needs.